Base code
Autonomous and teleop template.
MotorController class is for controlling motors with much more easer and
less code redunancies
Navigation class is for contolling the robot during autonomous
MotorNames enum holds all the names of the motors that are used by the
phones. If you want to add a motor or change a motor name then do it
here.